Fix GCC issues

+++ apps/constellation/main.cpp
@@ -201,8 +201,8 @@ struct CommandLineArguments
     if (args.bootstrap && args.token.size())
     {
       // determine what the P2P port is. This is either specified with the port parameter or
-      // explictly given via the manifest
-      uint16_t p2p_port = args.port + P2P_PORT_OFFSET;
+      // explicitly given via the manifest
+      uint16_t p2p_port = static_cast<uint16_t>(args.port + P2P_PORT_OFFSET);
 
       // if we have a valid manifest then we should respect the port configuration specified here
       // otherwise we default to the port specified
@@ -216,7 +216,7 @@ struct CommandLineArguments
         }
       }
 
-      // create the boostrap node
+      // create the bootstrap node
       bootstrap = std::make_unique<fetch::BootstrapMonitor>(
           prover.identity(), p2p_port, args.network_id, args.token, args.host_name);
 
@@ -294,7 +294,7 @@ struct CommandLineArguments
     manifest->AddService(ServiceIdentifier{ServiceType::HTTP}, Manifest::Entry{Uri{peer}});
 
     // register the P2P service
-    peer.Update(external_address, port + P2P_PORT_OFFSET);
+    peer.Update(external_address, static_cast<uint16_t>(port + P2P_PORT_OFFSET));
     manifest->AddService(ServiceIdentifier{ServiceType::P2P}, Manifest::Entry{Uri{peer}});
 
     // register all of the lanes (storage shards)
